% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/buffons_needle.R
\name{buffons_needle}
\alias{buffons_needle}
\title{Buffon's needle}
\usage{
buffons_needle(drops, w = 2, n = 1)
}
\arguments{
\item{drops}{The number of needle drops to simulate.}

\item{w}{The width of the boards or distance between lines.}

\item{n}{The length of the needle.}
}
\value{
A rational approximation of π, calculated as drops / hits.
}
\description{
This function simulates the mathematics problem, Buffon's needle,
proposed by Georges-Louis Leclerc, Comte de Buffon.
}
\examples{
buffons_needle(10000)
}

