% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/monty.R
\name{monty}
\alias{monty}
\title{Monty Hall problem}
\usage{
monty(n, shift = TRUE)
}
\arguments{
\item{n}{The number of games to simulate.}

\item{shift}{Whether to switch to a second door, a boolean value.}
}
\value{
A rational approximation of the probability that a player
will win a new car after switching their original door choice after
a non-winning door is revealed.
}
\description{
This function simulates the probability that a game show contestant,
searching for a new car behind a set of three closed doors, will
win the prize upon switching their original door choice after a non-
winning door is revealed. The problem was first proposed by Steve
Selvin.
}
\examples{
monty(10000)
}

