% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/petersburg.R
\name{petersburg}
\alias{petersburg}
\title{St. Petersburg paradox}
\usage{
petersburg(x, bet = 0)
}
\arguments{
\item{x}{The number of lotteries to simulate.}

\item{be}{The price at which the player plays the game.}
}
\value{
The average payout across all simulated lotteries.
}
\description{
This function simulates the average payout of a player in the lottery
game and consequent probability problem proposed by Nicolas
Bernoulli.
}
\examples{
petersburg(10000, bet = 15)
}

