% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/prisoners.R
\name{prisoners}
\alias{prisoners}
\title{100 prisoners problem}
\usage{
prisoners(sims, pris = 100, trys = 0.5)
}
\arguments{
\item{sims}{The number of survival situations to simulate.}

\item{pris}{The number of prisoners.}

\item{trys}{The ratio of attempts each prisoner gets relative to the
number of drawers.}
}
\value{
A rational approximation of the probability that a the group
of prisoners will find each of their own numbers.
}
\description{
This function simulates the probability that a group of 100
prisoners are able to find each of their own numbers in one of 100
drawers by following this algorithm:
}
\details{
1. A prisoner first opens the drawer with her own number.
2. If the drawer contains her number, she is finished.
3. Otherwise, the drawer contains another number and the prisoner
opens the drawer with that number next.
4. The prisoner repeats steps 2 and 3 until she finds her number
or has opened n / 2 drawers.

The problem was first proposed by Peter Bro Miltersen.
}
\examples{
prisoners(1000)
}

