\name{rcarlo}
\alias{rcarlo}
\docType{package}
\title{
Monte Carlo simulations
}
\description{
Miscellaneous exercises in Monte Carlo simulation.

This R package is a compilation of functions that model mathematical problems using repeated random sampling.
}
\details{
\tabular{ll}{
Package: \tab rcarlo\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-05-27\cr
License: \tab GPL-3\cr
}
}
\author{
Ben Cunningham <benjamin-cunningham@uiowa.edu>
}
\keyword{monte carlo}
\keyword{simulation}